requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.2.3
numpy==1.26.2
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.3.0
catboost==1.2.5
joblib==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.0.0
